changes in token and validate login and crud in currency changes
