+++ application/config/jwt.php
@@ -34,4 +34,4 @@ $config['token_header'] = 'authorization';
 | ( 1 Hour ) : 60 * 60     = 3600
 | ( 1 Minute ) : 60        = 60
 */
-$config['token_expire_time'] = 60;
+$config['token_expire_time'] = 7200;
