+++ index.php
@@ -86,7 +86,7 @@ switch (ENVIRONMENT)
 {
 	case 'development':
 		error_reporting(0);
-		ini_set('display_errors', 0);
+		ini_set('display_errors', 1);
 	break;
 
 	case 'testing':
